+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 ----------------------
 
 * Update :meth:`api.core.sort_alleles` method to also sort alleles by name for genes that do not use the star allele nomenclature (e.g. the DPYD gene).
+* Add new method :meth:`api.core.is_legit_allele`.
 
 0.7.0 (2021-10-23)
 ------------------